requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx==0.26.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
redis==5.0.1
cryptography==42.0.0
beautifulsoup4==4.12.3
lxml==5.1.0
bencode.py==4.0.0
